[CAN-30] refactor: 리뷰 반영해서 수정
